Moved package contents of sensai.geoanalytics to sensai.geoanalytics.geopandas to make room for other non-geopandas-dependent geoanalytics modules that shall be usable without requiring geopandas
