--- src/sensai/geoanalytics/__init__.py
@@ -1,1 +1,0 @@
-from . import coordinates, coordinate_clustering, coordinate_clustering_ground_truth, geometry
