+++ src/sensai/geoanalytics/geopandas/coordinate_clustering.py
@@ -4,12 +4,12 @@ import numpy as np
 from shapely.geometry import MultiPoint
 from typing import Callable, Union, Iterable
 
-from ..clustering.clustering_base import EuclideanClusterer
-from ..clustering import SkLearnEuclideanClusterer
-from ..clustering.sklearn_clustering import SkLearnClustererProtocol
-from ..util.cache import LoadSaveInterface
+from ...clustering.clustering_base import EuclideanClusterer
+from ...clustering import SkLearnEuclideanClusterer
+from ...clustering.sklearn_clustering import SkLearnClustererProtocol
+from ...util.cache import LoadSaveInterface
 from .coordinates import validateCoordinates, extractCoordinatesArray, TCoordinates, GeoDataFrameWrapper
-from ..util.profiling import timed
+from ...util.profiling import timed
 
 log = logging.getLogger(__name__)
