+++ tests/frameworks/geoanalytics/test_geoanalytics.py
@@ -1,3 +1,7 @@
-def test_geoanalytics():
-    import sensai.geoanalytics
+from sensai.util import markUsed
+
+
+def test_geoanalytics_geopandas():
+    import sensai.geoanalytics.geopandas
+    markUsed(sensai.geoanalytics.geopandas)
     assert True
